+++ EasyPost.Tests/Fixture.cs
@@ -51,7 +51,7 @@ namespace EasyPost.Tests._Utilities
             {
                 Dictionary<string, object> fixture = GetFixtureStructure().Pickups.Basic;
 
-                const string pickupDate = "2022-10-25";
+                const string pickupDate = "2023-01-26";
 
                 fixture.AddOrUpdate("min_datetime", pickupDate);
                 fixture.AddOrUpdate("max_datetime", pickupDate);