+++ EasyPost.Tests/ServicesTests/Beta/RateServiceTest.cs
@@ -1,0 +1,70 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using EasyPost.Models.API.Beta;
+using EasyPost.Tests._Utilities;
+using EasyPost.Tests._Utilities.Annotations;
+using EasyPost.Utilities;
+using EasyPost.Utilities.Annotations;
+using Xunit;
+
+namespace EasyPost.Tests.ServicesTests.Beta
+{
+    public class RateServiceTests : UnitTest
+    {
+        public RateServiceTests() : base("beta_rate_service")
+        {
+        }
+
+        #region Tests
+
+        #region Test CRUD Operations
+
+        [Fact]
+        [CrudOperations.Read]
+        [Testing.Function]
+        public async Task TestRetrieve()
+        {
+            UseVCR("retrieve");
+
+            Dictionary<string, object> shipmentData = Fixtures.BasicShipment;
+
+            List<StatelessRate> rates = await Client.Beta.Rate.RetrieveStatelessRates(shipmentData);
+
+            Assert.IsType<List<StatelessRate>>(rates);
+        }
+
+        #endregion
+
+        [Fact]
+        [Testing.Function]
+        public async Task TestGetLowestRateStaticFunction()
+        {
+            UseVCR("get_lowest_rate_static_function");
+
+            Dictionary<string, object> shipmentData = Fixtures.BasicShipment;
+
+            List<StatelessRate> rates = await Client.Beta.Rate.RetrieveStatelessRates(shipmentData);
+
+            StatelessRate lowestStatelessRate = Utilities.Rates.GetLowestStatelessRate(rates);
+
+            Assert.Equal("First", lowestStatelessRate.Service);
+        }
+
+        [Fact]
+        [Testing.Function]
+        public async Task TestGetLowestRateExtensionFunction()
+        {
+            UseVCR("get_lowest_rate_extension_function");
+
+            Dictionary<string, object> shipmentData = Fixtures.BasicShipment;
+
+            List<StatelessRate> rates = await Client.Beta.Rate.RetrieveStatelessRates(shipmentData);
+
+            StatelessRate lowestStatelessRate = rates.GetLowest();
+
+            Assert.Equal("First", lowestStatelessRate.Service);
+        }
+
+        #endregion
+    }
+}