+++ EasyPost/Services/RateService.cs
@@ -1,7 +1,7 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using EasyPost._base;
-using EasyPost.Calculation;
 using EasyPost.Models.API;
 using EasyPost.Utilities.Annotations;
 
@@ -29,6 +29,8 @@ namespace EasyPost.Services
 
         /// <summary>
         ///     Get the lowest rate from a list of rates.
+        ///
+        ///     Deprecated. Use <see cref="EasyPost.Utilities.Rates.GetLowestRate(IEnumerable{EasyPost.Models.API.Rate}, List{string}?, List{string}?, List{string}?, List{string}?)"/> instead.
         /// </summary>
         /// <param name="rates">List of rates to filter.</param>
         /// <param name="includeCarriers">Carriers to include in the filter.</param>
@@ -36,8 +38,9 @@ namespace EasyPost.Services
         /// <param name="excludeCarriers">Carriers to exclude in the filter.</param>
         /// <param name="excludeServices">Services to exclude in the filter.</param>
         /// <returns>Lowest EasyPost.Rate object instance.</returns>
+        [Obsolete("This method is deprecated. Please use EasyPost.Utilities.Rates.GetLowestRate() instead. This method will be removed in a future version.", false)]
 #pragma warning disable CA1822
-        public Rate GetLowestRate(IEnumerable<Rate> rates, List<string>? includeCarriers = null, List<string>? includeServices = null, List<string>? excludeCarriers = null, List<string>? excludeServices = null) => Rates.GetLowestObjectRate(rates, includeCarriers, includeServices, excludeCarriers, excludeServices);
+        public Rate GetLowestRate(IEnumerable<Rate> rates, List<string>? includeCarriers = null, List<string>? includeServices = null, List<string>? excludeCarriers = null, List<string>? excludeServices = null) => Utilities.Rates.GetLowestRate(rates, includeCarriers, includeServices, excludeCarriers, excludeServices);
 #pragma warning restore CA1822
     }
 }